requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
apify==2.5.0
apify_client==1.9.2
apify_fingerprint_datapoints==0.0.2
apify_shared==1.3.2
astrapy==2.0.1
attrs==25.3.0
Brotli==1.1.0
browserforge==1.2.3
cachetools==5.5.2
cassandra-driver==3.29.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crawlee==0.6.5
cryptography==44.0.2
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
eval_type_backport==0.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.1
geomet==0.2.1.post1
groq==0.20.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.49
langchain-text-splitters==0.3.7
langgraph==0.3.21
langgraph-checkpoint==2.0.23
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.60
langsmith==0.3.19
lazy-object-proxy==1.10.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.4
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pillow==11.1.0
propcache==0.3.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.6.1
pydantic_core==2.33.0
pyee==13.0.0
Pygments==2.19.1
pymongo==4.11.3
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.1
safetensors==0.5.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcollections==2.1.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.40
starlette==0.46.1
sympy==1.13.1
tenacity==9.0.0
tldextract==5.1.3
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.50.3
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uuid6==2024.7.10
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
